Replace ' ' with \s in regex pattern if EXTENDED flag is set.
PCRE2 skips ' ' by default if the EXTENDED flag is set.
POSIX regex doesn't. I assume that a lot of folks rely on POSIX behavior
in this regard, and it is a lot of work to change it.
Also, it is more convenient to type a simple space in the command line instead of a \s